+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# [0.0.1-alpha.3] - 2021-03-02
+
+## Added
+- options for `LoginWithPopup` (`PopupLoginOptions` and `PopupConfigOptions`)
+- `BaseLoginOptions`
+- scope for login in example
+
 # [0.0.1-alpha.20210224] - 2021-02-24
 
 ## Added